KUDU-219: Initialize correctly_, prevent underflow
This revision fixes KUDU-219 by initializing num_flush_pins_ to 0
the other constructor. Additionally, num_flush_pins_ is changed to
a signed integer CHECK_GE(num_flush_pins_, 0) is added in few places.

Reviewed-on: http://gerrit.ent.cloudera.com:8080/2171
Tested-by: jenkins

+++ src/server/metadata.h
@@ -227,7 +227,7 @@ class TabletMetadata {
 
   // If this counter is > 0 then Flush() will not write any data to
   // disk.
-  size_t num_flush_pins_;
+  int32_t num_flush_pins_;
 
   // Set if Flush() is called when num_flush_pins_ is > 0; if true,
   // then next UnPinFlush will call Flush() again to ensure the
